+++ tests/spadl/test_statsbomb.py
@@ -1,6 +1,5 @@
 import os
 
-import pandas as pd
 import pytest
 
 from socceraction.spadl import config as spadl
@@ -53,115 +52,38 @@ class TestStatsBombLoader:
 class TestSpadlConvertor:
     def setup_method(self):
         data_dir = os.path.join(os.path.dirname(__file__), os.pardir, 'data', 'statsbomb', 'raw')
-        SBL = sb.StatsBombLoader(root=data_dir, getter='local')
-        self.events = SBL.events(7584)
+        self.SBL = sb.StatsBombLoader(root=data_dir, getter='local')
+        # https://raw.githubusercontent.com/statsbomb/open-data/master/data/events/7584.json
+        self.id_bel = 782
+        self.events_japbel = self.SBL.events(7584)
 
     def test_extract_player_games(self):
-        df_player_games = sb.extract_player_games(self.events)
+        df_player_games = sb.extract_player_games(self.events_japbel)
         assert len(df_player_games) == 26
         assert len(df_player_games.player_name.unique()) == 26
         assert set(df_player_games.team_name) == {'Belgium', 'Japan'}
         assert df_player_games.minutes_played.sum() == 22 * 95
 
     def test_convert_to_actions(self):
-        df_actions = sb.convert_to_actions(self.events, 782)
+        df_actions = sb.convert_to_actions(self.events_japbel, 782)
         assert len(df_actions) > 0
         SPADLSchema.validate(df_actions)
         assert (df_actions.game_id == 7584).all()
         assert ((df_actions.team_id == 782) | (df_actions.team_id == 778)).all()
 
     def test_convert_start_location(self):
-        event = pd.DataFrame(
-            [
-                {
-                    'event_id': 'a1b55211-a292-4294-887b-5385cc3c5705',
-                    'index': 5,
-                    'period_id': 1,
-                    'timestamp': '00:00:00.920',
-                    'minute': 0,
-                    'second': 0,
-                    'type_id': 30,
-                    'type_name': 'Pass',
-                    'possession': 2,
-                    'possession_team_id': 782,
-                    'possession_team_name': 'Belgium',
-                    'play_pattern_id': 9,
-                    'play_pattern_name': 'From Kick Off',
-                    'team_id': 782,
-                    'team_name': 'Belgium',
-                    'duration': 0.973,
-                    'extra': {
-                        'pass': {
-                            'recipient': {'id': 5642, 'name': 'Axel Witsel'},
-                            'length': 12.369317,
-                            'angle': 2.896614,
-                            'height': {'id': 1, 'name': 'Ground Pass'},
-                            'type': {'id': 65, 'name': 'Kick Off'},
-                            'body_part': {'id': 40, 'name': 'Right Foot'},
-                        }
-                    },
-                    'related_events': ['051449c5-e183-46f9-965d-1d8f00f017cb'],
-                    'player_id': 3289.0,
-                    'player_name': 'Romelu Lukaku Menama',
-                    'position_id': 23.0,
-                    'position_name': 'Center Forward',
-                    'location': [61.0, 40.0],
-                    'under_pressure': None,
-                    'counterpress': None,
-                    'game_id': 7584,
-                }
-            ]
-        )
-        action = sb.convert_to_actions(event, 782).iloc[0]
+        event = self.events_japbel[
+            self.events_japbel.event_id == 'a1b55211-a292-4294-887b-5385cc3c5705'
+        ]
+        action = sb.convert_to_actions(event, self.id_bel).iloc[0]
         assert action['start_x'] == ((61.0 - 1) / 119) * spadl.field_length
         assert action['start_y'] == 68 - ((40.0 - 1) / 79) * spadl.field_width
-        assert action['end_x'] == action['start_x']
-        assert action['end_y'] == action['start_y']
 
     def test_convert_end_location(self):
-        event = pd.DataFrame(
-            [
-                {
-                    'event_id': 'a1b55211-a292-4294-887b-5385cc3c5705',
-                    'index': 5,
-                    'period_id': 1,
-                    'timestamp': '00:00:00.920',
-                    'minute': 0,
-                    'second': 0,
-                    'type_id': 30,
-                    'type_name': 'Pass',
-                    'possession': 2,
-                    'possession_team_id': 782,
-                    'possession_team_name': 'Belgium',
-                    'play_pattern_id': 9,
-                    'play_pattern_name': 'From Kick Off',
-                    'team_id': 782,
-                    'team_name': 'Belgium',
-                    'duration': 0.973,
-                    'extra': {
-                        'pass': {
-                            'recipient': {'id': 5642, 'name': 'Axel Witsel'},
-                            'length': 12.369317,
-                            'angle': 2.896614,
-                            'height': {'id': 1, 'name': 'Ground Pass'},
-                            'end_location': [49.0, 43.0],
-                            'type': {'id': 65, 'name': 'Kick Off'},
-                            'body_part': {'id': 40, 'name': 'Right Foot'},
-                        }
-                    },
-                    'related_events': ['051449c5-e183-46f9-965d-1d8f00f017cb'],
-                    'player_id': 3289.0,
-                    'player_name': 'Romelu Lukaku Menama',
-                    'position_id': 23.0,
-                    'position_name': 'Center Forward',
-                    'location': [61.0, 40.0],
-                    'under_pressure': None,
-                    'counterpress': None,
-                    'game_id': 7584,
-                }
-            ]
-        )
-        action = sb.convert_to_actions(event, 782).iloc[0]
+        event = self.events_japbel[
+            self.events_japbel.event_id == 'a1b55211-a292-4294-887b-5385cc3c5705'
+        ]
+        action = sb.convert_to_actions(event, self.id_bel).iloc[0]
         assert action['end_x'] == ((49.0 - 1) / 119) * spadl.field_length
         assert action['end_y'] == 68 - ((43.0 - 1) / 79) * spadl.field_width
 
@@ -178,49 +100,14 @@ class TestSpadlConvertor:
         ],
     )
     def test_convert_time(self, period, timestamp, minute, second):
-        event = pd.DataFrame(
-            [
-                {
-                    'event_id': 'a1b55211-a292-4294-887b-5385cc3c5705',
-                    'index': 5,
-                    'period_id': period,
-                    'timestamp': timestamp,
-                    'minute': minute,
-                    'second': second,
-                    'type_id': 30,
-                    'type_name': 'Pass',
-                    'possession': 2,
-                    'possession_team_id': 782,
-                    'possession_team_name': 'Belgium',
-                    'play_pattern_id': 9,
-                    'play_pattern_name': 'From Kick Off',
-                    'team_id': 782,
-                    'team_name': 'Belgium',
-                    'duration': 0.973,
-                    'extra': {
-                        'pass': {
-                            'recipient': {'id': 5642, 'name': 'Axel Witsel'},
-                            'length': 12.369317,
-                            'angle': 2.896614,
-                            'height': {'id': 1, 'name': 'Ground Pass'},
-                            'end_location': [49.0, 43.0],
-                            'type': {'id': 65, 'name': 'Kick Off'},
-                            'body_part': {'id': 40, 'name': 'Right Foot'},
-                        }
-                    },
-                    'related_events': ['051449c5-e183-46f9-965d-1d8f00f017cb'],
-                    'player_id': 3289.0,
-                    'player_name': 'Romelu Lukaku Menama',
-                    'position_id': 23.0,
-                    'position_name': 'Center Forward',
-                    'location': [61.0, 40.0],
-                    'under_pressure': None,
-                    'counterpress': None,
-                    'game_id': 7584,
-                }
-            ]
-        )
-        action = sb.convert_to_actions(event, 782).iloc[0]
+        event = self.events_japbel[
+            self.events_japbel.event_id == 'a1b55211-a292-4294-887b-5385cc3c5705'
+        ].copy()
+        event['period_id'] = period
+        event['timestamp'] = timestamp
+        event['minute'] = minute
+        event['second'] = second
+        action = sb.convert_to_actions(event, self.id_bel).iloc[0]
         assert action['period_id'] == period
         assert (
             action['time_seconds']
@@ -233,51 +120,28 @@ class TestSpadlConvertor:
         )
 
     def test_convert_pass(self):
-        pass_event = pd.DataFrame(
-            [
-                {
-                    'event_id': 'a1b55211-a292-4294-887b-5385cc3c5705',
-                    'index': 5,
-                    'period_id': 1,
-                    'timestamp': '00:00:00.920',
-                    'minute': 0,
-                    'second': 0,
-                    'type_id': 30,
-                    'type_name': 'Pass',
-                    'possession': 2,
-                    'possession_team_id': 782,
-                    'possession_team_name': 'Belgium',
-                    'play_pattern_id': 9,
-                    'play_pattern_name': 'From Kick Off',
-                    'team_id': 782,
-                    'team_name': 'Belgium',
-                    'duration': 0.973,
-                    'extra': {
-                        'pass': {
-                            'recipient': {'id': 5642, 'name': 'Axel Witsel'},
-                            'length': 12.369317,
-                            'angle': 2.896614,
-                            'height': {'id': 1, 'name': 'Ground Pass'},
-                            'end_location': [49.0, 43.0],
-                            'type': {'id': 65, 'name': 'Kick Off'},
-                            'body_part': {'id': 40, 'name': 'Right Foot'},
-                        }
-                    },
-                    'related_events': ['051449c5-e183-46f9-965d-1d8f00f017cb'],
-                    'player_id': 3289.0,
-                    'player_name': 'Romelu Lukaku Menama',
-                    'position_id': 23.0,
-                    'position_name': 'Center Forward',
-                    'location': [61.0, 40.0],
-                    'under_pressure': None,
-                    'counterpress': None,
-                    'game_id': 7584,
-                }
-            ]
-        )
-        pass_action = sb.convert_to_actions(pass_event, 782).iloc[0]
+        pass_event = self.events_japbel[
+            self.events_japbel.event_id == 'a1b55211-a292-4294-887b-5385cc3c5705'
+        ]
+        pass_action = sb.convert_to_actions(pass_event, self.id_bel).iloc[0]
         assert pass_action['team_id'] == 782
         assert pass_action['player_id'] == 3289
         assert pass_action['type_id'] == spadl.actiontypes.index('pass')
         assert pass_action['result_id'] == spadl.results.index('success')
         assert pass_action['bodypart_id'] == spadl.bodyparts.index('foot')
+
+    def test_convert_own_goal(self):
+        events_morira = self.SBL.events(7577)
+        own_goal_for_event = events_morira[
+            events_morira.event_id == '8981bc58-6041-4b78-95c5-ebe9677ca379'
+        ]
+        own_goal_for_actions = sb.convert_to_actions(own_goal_for_event, 797)
+        assert len(own_goal_for_actions) == 0
+        own_goal_against_event = events_morira[
+            events_morira.event_id == 'cef0fcb6-28d0-49d7-8f93-4a0aef28001a'
+        ]
+        own_goal_against_actions = sb.convert_to_actions(own_goal_against_event, 797)
+        assert len(own_goal_against_actions) == 1
+        assert own_goal_against_actions.iloc[0]['type_id'] == spadl.actiontypes.index('bad_touch')
+        assert own_goal_against_actions.iloc[0]['result_id'] == spadl.results.index('owngoal')
+        assert own_goal_against_actions.iloc[0]['bodypart_id'] == spadl.bodyparts.index('foot')